Changes to be committed: 	modified:   ASCII3D.depend 	modified:   ASCII3D.layout 	modified:   bin/Debug/ASCII3D.exe 	modified:   headers/console.h 	modified:   obj/Debug/source/Player.o 	modified:   obj/Debug/source/console.o 	modified:   obj/Debug/source/global.o 	modified:   obj/Debug/source/input.o 	modified:   obj/Debug/source/main.o 	modified:   source/Console.cpp 	modified:   source/Player.cpp 	modified:   source/global.cpp Basicamente he arreglado lo de las texturas culeadas, ahora estoy intentando hacerlo independiente del tamaño de la pantalla.

+++ headers/console.h
@@ -22,7 +22,7 @@ class Console {
     int drawString(char *text, int aSize, int xPos, int yPos, bool specify = false, short int attribute = 0);
     int refreshCursor();
     int drawPixel(int xPos, int yPos, int attribute, int intestity = 4);
-    int drawLine(int lSize, int texIndex, int lPos, float texPos);
+    int drawLine(float lSize, int texIndex, int lPos, float texPos);
 
 
     int sWidth;
